Update README.md
Dear Viewer,

This code is solely my proprietary and it is not connected with any company or organizations.

I request companies and organizations not use my link for yours work sample

Regards

Shankar Kamble

Using the MEAN stack, create a web page that shuffles a standard deck of 52 cards, and then deals 

them in a matrix of 13 columns by 4 rows. Calculate the percentage that the deck is in the correct order 

and store the deck configuration and the percentage in the database. Display an average percentage 

across all historical percentages.

For every card that is in the correct suited row, add one point

For every card that is in the correct column, add one point

Max points = 104

This configuration would give you 104 points = 100% correct order

MongoDB: use mongo as the persistent datastore

Express.js: use express as the web server to handle routing the calls

Angular.js: use angular framework as the front-end templating engine

Node.js: use node to handle all the back-end business logic and database access